requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pydantic==1.10.4
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
